+++ src/app.routing.ts
@@ -1,8 +1,7 @@
 import { Routes } from "@angular/router";
-import {
-  PathNotFoundComponentComponent
-} from "./app/core/components/path-not-found-component/path-not-found-component.component";
 import { AdminGuard, IsCartEmptyGuardGuard } from "./app/core/guards";
+import { PathNotFoundComponentComponent, SettingsComponent } from "./app/core/components";
+
 
 
 export const APP_ROUTES: Routes = [
@@ -24,6 +23,10 @@ export const APP_ROUTES: Routes = [
     path: 'order',
     canActivate: [IsCartEmptyGuardGuard],
     loadComponent: () => import('./app/core/components/process-order/process-order.component').then(c => c.ProcessOrderComponent)
+  },
+  {
+    path: 'settings',
+    component: SettingsComponent
   },
   {
     path: '**',