+++ src/app/core/helpers/constants.ts
@@ -1,4 +1,5 @@
 import { SortOptions } from "../models/constant.model";
+import {ISettings, Methods} from "../models/settings.model";
 
 export const characters ='ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
 
@@ -7,3 +8,26 @@ export const sortOptions: SortOptions = {
   count: 'count',
   name: 'title'
 }
+
+export const defaultSettings: ISettings = {
+  log: {
+    defaultLogValues: [Methods.Get]
+  },
+  sort: {
+    defaultLogValues: {
+      order: "count",
+      direction: false
+    }
+  }
+};
+
+
+const API_URL = "http://localhost:3000";
+export const productsUrl = `${API_URL}/products`;
+export const cartUrl = `${API_URL}/cart`;
+
+export const settingsUrl = 'assets/app-settings.json'
+
+export const cartLocalStorageName = 'cart';
+export const logSettingStorageName = 'logSetting';
+export const sortSettingStorageName = 'sortSetting';