+++ src/main.ts
@@ -1,22 +1,24 @@
 import { bootstrapApplication } from "@angular/platform-browser";
-import {DEFAULT_CURRENCY_CODE, importProvidersFrom} from "@angular/core";
+import { DEFAULT_CURRENCY_CODE, importProvidersFrom } from "@angular/core";
 import { provideAnimations } from "@angular/platform-browser/animations";
+import { RouterModule } from "@angular/router";
+import { HttpClientModule } from "@angular/common/http";
 
 import { AppComponent} from "./app/app.component";
-import { RouterModule } from "@angular/router";
 import { APP_ROUTES } from "./app.routing";
-import {ProductsModule} from "./app/products/products.module";
+import { ProductsModule } from "./app/products/products.module";
+import { httpInterceptorProviders } from "./app/core/interceptors";
 
 
 
 bootstrapApplication(AppComponent, {
   providers: [
+    httpInterceptorProviders,
     provideAnimations(),
     {
       provide: DEFAULT_CURRENCY_CODE,
       useValue: 'CAD'
     },
-    importProvidersFrom(ProductsModule),
-    importProvidersFrom(RouterModule.forRoot(APP_ROUTES)),
+    importProvidersFrom([ProductsModule, HttpClientModule, RouterModule.forRoot(APP_ROUTES)]),
   ],
 })